Move binary-only dependencies under a feature
