+++ Cargo.toml
@@ -17,23 +17,25 @@ description = "Creates a book from markdown files"
 
 [dependencies]
 anyhow = "1.0.28"
-chrono = "0.4"
-clap = { version = "3.0", features = ["cargo"] }
-clap_complete = "3.0"
 once_cell = "1"
-env_logger = "0.9.0"
-handlebars = "4.0"
 log = "0.4"
 memchr = "2.0"
-opener = "0.5"
 pulldown-cmark = { version = "0.9.1", default-features = false }
 regex = "1.5.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 shlex = "1"
-tempfile = "3.0"
 toml = "0.5.1"
-topological-sort = "0.1.0"
+
+# Bin feature
+chrono = { version = "0.4", optional = true }
+clap = { version = "3.0", features = ["cargo"], optional = true }
+clap_complete = { version = "3.0", optional = true }
+env_logger = { version = "0.9.0", optional = true }
+handlebars = { version = "4.0", optional = true }
+opener = { version = "0.5", optional = true }
+tempfile = { version = "3.0", optional = true }
+topological-sort = { version = "0.1.0", optional = true }
 
 # Watch feature
 notify = { version = "4.0", optional = true }
@@ -49,6 +51,8 @@ elasticlunr-rs = { version = "3.0.0", optional = true }
 ammonia = { version = "3", optional = true }
 
 [dev-dependencies]
+# Make "full" a required feature for tests
+mdbook = {path = ".", features = ["full"]}
 assert_cmd = "1"
 predicates = "2"
 select = "0.5"
@@ -57,10 +61,11 @@ pretty_assertions = "1.2.1"
 walkdir = "2.0"
 
 [features]
-default = ["watch", "serve", "search"]
+default = ["watch", "serve", "search", "full"]
 watch = ["notify", "gitignore"]
 serve = ["futures-util", "tokio", "warp"]
 search = ["elasticlunr-rs", "ammonia"]
+full = ["chrono", "clap", "clap_complete", "env_logger", "handlebars", "opener", "tempfile", "topological-sort"]
 
 [[bin]]
 doc = false
